+++ packages/mapsforge_flutter/lib/src/label/label_view.dart
@@ -4,6 +4,7 @@ import 'package:mapsforge_flutter/src/label/label_job_queue.dart';
 import 'package:mapsforge_flutter/src/label/label_painter.dart';
 import 'package:mapsforge_flutter/src/transform_widget.dart';
 import 'package:mapsforge_flutter_core/utils.dart';
+import 'package:mapsforge_flutter_renderer/offline_renderer.dart';
 
 /// A view to display the labels. The view updates itself whenever the [MapPosition] changes and new labels are available.
 /// Labels are drawn separately so they keep orientation.
@@ -12,12 +13,15 @@ import 'package:mapsforge_flutter_core/utils.dart';
 class LabelView extends StatefulWidget {
   final MapModel mapModel;
 
+  final Renderer renderer;
+
   /// Hide labels when zoomlevel < minLabelZoom. Null = feature disabled.
   final int? minLabelZoom;
 
   const LabelView({
     super.key,
     required this.mapModel,
+    required this.renderer,
     this.minLabelZoom, // e.g. pass 10 to suppress under zoom 10
   });
 
@@ -33,7 +37,7 @@ class _LabelViewState extends State<LabelView> {
   @override
   void initState() {
     super.initState();
-    jobQueue = LabelJobQueue(mapModel: widget.mapModel);
+    jobQueue = LabelJobQueue(mapModel: widget.mapModel, renderer: widget.renderer);
   }
 
   @override